backend/requirements.txt
# Moonkeep Elite - Backend Dependencies
# Core API
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
websockets>=11.0

# Authentication & Security
PyJWT>=2.8.0
bcrypt>=4.0.0
slowapi>=0.1.9

# Networking & Packet Manipulation
scapy>=2.5.0
requests>=2.28.0

# Recon-ng Dependencies
pyyaml>=6.0
dnspython>=2.3.0
lxml>=4.9.0
mechanize>=0.4.0
flask>=3.0.0
flask-restful>=0.3.10
flasgger>=0.9.7
dicttoxml>=1.7.0
xlsxwriter>=3.0.0
unicodecsv>=0.14.0
rq>=1.0.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
